raft_topology, tablets: Drain tablets in parallel with other topology operations
Allows bootstrap to execute while tablets are being drained on decommission.
Allows multiple decommission/removenode to happen in parallel.

Flow of decommission/removenode request:
  1) pending, has tablet replicas on target node
  2) not pending, paused (node::is_drained() == true)
  3) pending, no tablet replicas on target node
  4) executing, node is in transition
  5) done

Request is unpaused when topology state machine is idle, which makes
the request pending and ready to be scheduled. Tablet load balancing
will yield the state machine as soon as some request can be unpaused.

TODO:
 1) Failure propagation
 2) API for canceling requests
 3) Unit tests